Python/Scripts/pytesseract/requirements.txt
appdirs==1.4.3
olefile==0.44
packaging==16.8
Pillow==4.1.1
pyparsing==2.2.0
pytesseract==0.1.7
six==1.10.0
